+++ src/azure-cli/HISTORY.rst
@@ -5,7 +5,7 @@ Release History
 
 **Compute**
 
-* vmss create: Add --scale-in-policy, which decides which virtual machines are chosen for removal when a VMSS is scaled-in
+* vmss create/update: Add --scale-in-policy, which decides which virtual machines are chosen for removal when a VMSS is scaled-in
 
 **Install**